+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   default. If the browser sends an incorrect packet rate, this will be logged
   as an error in the console.
 - Add a warning log in `InMemoryJSONEventBuffer`'s `send` function when retrying starts.
+- Update NScale uplink policy to scale down resolution based on number of remote videos
 
 ### Changed
 - Update `InMemoryJSONEventBuffer` to retry with backoff.